+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/EditReplacePatchDeletable.java
@@ -17,8 +17,10 @@
 package io.fabric8.kubernetes.client.dsl;
 
 import io.fabric8.kubernetes.client.GracePeriodConfigurable;
+import io.fabric8.kubernetes.client.PropagationPolicyConfigurable;
 
 public interface EditReplacePatchDeletable<I, T, D, B> extends EditReplacePatchable<I, T, D>, Deletable<B>,
-                                                          GracePeriodConfigurable<Deletable<B>>
+                                                          GracePeriodConfigurable<Deletable<B>>,
+                                                          PropagationPolicyConfigurable<Deletable<B>>
 {
 }