+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 ## CHANGELOG
 
 ### 4.4-SNAPSHOT
+
 #### Bugs
   * Fix #1730: Fix failing build on jdk11
   * Fix #1634: Cascade delete can't be overriden
 
 #### Improvements
+  * Allow user to set a propagation policy on deletion
 
 #### Dependency Upgrade